Feat: friendList 뷰모델 구현
